+++ CHANGELOG.md
@@ -96,6 +96,7 @@
 - Crypto.com parser: added additional transaction kinds. ([#285](https://github.com/BittyTax/BittyTax/issues/285))
 - KuCoin parser: make direction field case-insensitive. ([#189](https://github.com/BittyTax/BittyTax/issues/189))
 - BlockFi parser: added "Referral Bonus" transaction type. ([#159](https://github.com/BittyTax/BittyTax/issues/159))
+- Kraken parser: use Ledgers export for trades.
 
 ## Version [0.5.0] Beta (2021-11-11)
 Important:-